+++ src/components/bluepad32/uni_property.c
@@ -36,11 +36,13 @@ const char* UNI_PROPERTY_KEY_MOUSE_SCALE = "bp.mouse.scale";
 // TODO: Move them to the Unijoysticle file.
 // Keep them sorted
 const char* UNI_PROPERTY_KEY_UNI_AUTOFIRE_CPS = "bp.uni.autofire";
+const char* UNI_PROPERTY_KEY_UNI_BB_FIRE_THRESHOLD = "bp.uni.bb_fire";
+const char* UNI_PROPERTY_KEY_UNI_BB_MOVE_THRESHOLD = "bp.uni.bb_move";
+const char* UNI_PROPERTY_KEY_UNI_C64_POT_MODE = "bp.uni.c64pot";
 const char* UNI_PROPERTY_KEY_UNI_MODEL = "bp.uni.model";
 const char* UNI_PROPERTY_KEY_UNI_MOUSE_EMULATION = "bp.uni.mouseemu";
 const char* UNI_PROPERTY_KEY_UNI_SERIAL_NUMBER = "bp.uni.serial";
 const char* UNI_PROPERTY_KEY_UNI_VENDOR = "bp.uni.vendor";
-const char* UNI_PROPERTY_KEY_UNI_C64_POT_MODE = "bp.uni.c64pot";
 
 // TODO: Implement "property interface" instead of doing #ifdef